Extract dynamo stream to share it with the event export
